Book Ticket
Create an API where you can submit a purchase for a ticket. Details included in the receipt are: a) From, To, User , price paid. (i) User should include first and last name, email address
The user is allocated a seat in the train. Assume the train has only 2 sections, section A and section B.
An API that shows the details of the receipt for the user
An API that lets you view the users and seat they are allocated by the requested section
An API to remove a user from the train
An API to modify a user's seat